Flask App/requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
comtypes==1.1.14
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.1
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.2.2
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
gast==0.4.0
google==3.0.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.21.0
keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.0.4
mlxtend==0.22.0
mongoose==0.0.1
MouseInfo==0.1.3
multidict==6.0.4
mysql-connector-python==8.0.33
nbclassic==0.5.4
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.0
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytweening==1.0.4
pytz==2023.3
pywhatkit==5.4
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.1
QtPy==2.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wikipedia==1.4.0
wrapt==1.14.1
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
